pyproject.toml
[tool.poetry]
name = "qa-quru-3-20-mobile"
version = "0.1.0"
description = ""
authors = ["KseniiaTr <ks31993@mail.ru>"]
readme = "README.md"
packages = [{include = "qa_quru_3_20_mobile"}]

[tool.poetry.dependencies]
python = "^3.11"
Appium-Python-Client = "^2.6.1"
selene = {version = "2.0.0-beta.8", allow-prereleases = true}
pytest = "^7.1.3"
pytest-xdist = "^2.5.0"
allure-pytest = "^2.10.0"
requests = "^2.28.1"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
